+++ src/axom/inlet/Proxy.hpp
@@ -76,12 +76,12 @@ public:
    * or T{}
    *******************************************************************************
    */
-  template <typename T,
-            typename SFINAE =
-              typename std::enable_if<detail::is_inlet_primitive<T>::value ||
-                                      detail::is_inlet_primitive_array<T>::value ||
-                                      detail::is_inlet_primitive_dict<T>::value ||
-                                      detail::is_std_function<T>::value>::type>
+  template <
+    typename T,
+    typename SFINAE = typename std::enable_if<
+      detail::is_inlet_primitive<T>::value || detail::is_inlet_primitive_array<T>::value ||
+      detail::is_inlet_primitive_dict<T>::value || detail::is_std_function<T>::value ||
+      detail::is_primitive_std_vector<T>::value>::type>
   operator T() const
   {
     return get<T>();